+++ docs/pages/release_notes.md
@@ -65,6 +65,12 @@ Since this release, PMD will also expose any getter returning a collection of an
   a build with errors will now fail and with that parameter, the previous behavior can be restored.
   This parameter is available for both PMD and CPD.
 
+##### Ant
+
+* CPDTask has a new parameter `failOnError`. In controls, whether to fail the build if any recoverable errors occurred.
+  By default, the build will fail. CPD will still create a report with all detected duplications, but the report might
+  be incomplete.
+
 #### Deprecated API
 
 * pmd-java